// DXC_Frontend/src/api/models/hotelWithImagesDto.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * DXC_Core.API
 * OpenAPI spec version: 1.0
 */
import type { HotelImageDto } from './hotelImageDto'

export interface HotelWithImagesDto {
  publicId?: string
  /** @nullable */
  name?: string | null
  /** @nullable */
  description?: string | null
  /** @nullable */
  address?: string | null
  /** @nullable */
  starRating?: number | null
  /** @nullable */
  phoneNumber?: string | null
  /** @nullable */
  email?: string | null
  /** @nullable */
  website?: string | null
  /** @nullable */
  operatingHours?: string | null
  /** @nullable */
  latitude?: number | null
  /** @nullable */
  longitude?: number | null
  /** @nullable */
  vr360Link?: string | null
  /** @nullable */
  priceFrom?: number | null
  /** @nullable */
  priceFromCurrency?: string | null
  thuTu?: number
  isActive?: boolean
  createdAt?: string
  /** @nullable */
  updatedAt?: string | null
  /** @nullable */
  images?: HotelImageDto[] | null
}